requirements.txt
streamlit
pandas
openpyxl
nest_asyncio
beautifulsoup4
newspaper3k
playwright
playwright_stealth
openai
python-dotenv
requests
lxml_html_clean
